+++ packages/check/src/main.rs
@@ -10,7 +10,8 @@ use colored::Colorize;
 use cosmwasm_vm::capabilities_from_csv;
 use cosmwasm_vm::internals::{check_wasm, compile};
 
-const DEFAULT_AVAILABLE_CAPABILITIES: &str = "iterator,staking,stargate,cosmwasm_1_1,cosmwasm_1_2";
+const DEFAULT_AVAILABLE_CAPABILITIES: &str =
+    "iterator,staking,stargate,cosmwasm_1_1,cosmwasm_1_2,cosmwasm_1_3";
 
 pub fn main() {
     let matches = Command::new("Contract checking")